Restructured the file structure and created user file
