--- advert_details.py
@@ -1,13 +1,0 @@
-from fastapi import HTTPException, status
-from bson.objectid import ObjectId
-from db import adverts_collection
-from utils import replace_advert_id
-
-
-def get_advert_by_id(advert_id):
-    if not ObjectId.is_valid(advert_id):
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Ad not found")
-
-    advert = adverts_collection.find_one({"_id": ObjectId(advert_id)})
-    return {"data": replace_advert_id(advert)}
