--- delete.py
@@ -1,18 +1,0 @@
-from fastapi import HTTPException, status
-from bson.objectid import ObjectId
-from db import adverts_collection
-
-
-def delete_advert(advert_id):
-    if not ObjectId.is_valid(advert_id):
-        raise HTTPException(
-            status.HTTP_422_UNPROCESSABLE_ENTITY, "Invalid mongo id received")
-    # Delete event from database
-    delete_result = adverts_collection.delete_one(
-        filter={"_id": ObjectId(advert_id)})
-    if not delete_result:
-        raise HTTPException(
-            status.HTTP_422_UNPROCESSABLE_ENTITY, "Invalid mongo id received")
-    
-    return {"message": f"Advert with id {advert_id} has been deleted successfully."}
-
